+++ src/Machina/Types/Vector.cs
@@ -36,6 +36,13 @@ namespace Machina
         /// </summary>
         public double Z { get; internal set; }
 
+        public static bool AreSimilar(Vector a, Vector b, double epsilon = 0.000001)
+        {
+            return Math.Abs(a.X - b.X) < EPSILON2
+                   && Math.Abs(a.Y - b.Y) < EPSILON2
+                   && Math.Abs(a.Z - b.Z) < EPSILON2;
+        }
+
         /// <summary>
         /// Test if this Vector is approximately equal to another.
         /// </summary>